Refactor for integration test isolation
